move methods where they make more sense
