+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/QueryInfo.java
@@ -12,38 +12,38 @@
  *******************************************************************************/
 package io.openliberty.data.internal.persistence;
 
+import static io.openliberty.data.internal.persistence.Util.SORT_PARAM_TYPES;
+import static io.openliberty.data.internal.persistence.Util.SPECIAL_PARAM_TYPES;
+import static io.openliberty.data.internal.persistence.Util.lifeCycleReturnTypes;
 import static io.openliberty.data.internal.persistence.cdi.DataExtension.exc;
 import static jakarta.data.repository.By.ID;
 
 import java.lang.annotation.Annotation;
+import java.lang.reflect.Array;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Member;
 import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
 import java.lang.reflect.Parameter;
 import java.lang.reflect.RecordComponent;
 import java.math.BigDecimal;
 import java.math.BigInteger;
-import java.time.Instant;
-import java.time.LocalDateTime;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.Iterator;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Map.Entry;
 import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 import java.util.stream.BaseStream;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import com.ibm.websphere.ras.Tr;
@@ -57,6 +57,8 @@ import jakarta.data.Order;
 import jakarta.data.Sort;
 import jakarta.data.exceptions.DataException;
 import jakarta.data.exceptions.MappingException;
+import jakarta.data.exceptions.NonUniqueResultException;
+import jakarta.data.exceptions.OptimisticLockingFailureException;
 import jakarta.data.page.CursoredPage;
 import jakarta.data.page.Page;
 import jakarta.data.page.PageRequest;
@@ -69,6 +71,9 @@ import jakarta.data.repository.Param;
 import jakarta.data.repository.Query;
 import jakarta.data.repository.Save;
 import jakarta.data.repository.Update;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.LockModeType;
+import jakarta.persistence.TypedQuery;
 
 /**
  * Query information.
@@ -77,14 +82,15 @@ public class QueryInfo {
     private static final TraceComponent tc = Tr.register(QueryInfo.class);
 
     public static enum Type {
-        COUNT, DELETE, DELETE_WITH_ENTITY_PARAM, EXISTS, FIND, FIND_AND_DELETE, INSERT, SAVE, RESOURCE_ACCESS,
+        COUNT,
+        DELETE, DELETE_WITH_ENTITY_PARAM,
+        EXISTS,
+        FIND, FIND_AND_DELETE,
+        INSERT,
+        SAVE,
+        RESOURCE_ACCESS,
         UPDATE, UPDATE_WITH_ENTITY_PARAM, UPDATE_WITH_ENTITY_PARAM_AND_RESULT
     }
-
-    /**
-     * Commonly used result types that are not entities.
-     */
-    private static final Set<Class<?>> NON_ENTITY_RESULT_TYPES = new HashSet<>();
 
     /**
      * Indicates the repository method has no Sort, Sort[], or Order parameters
@@ -105,82 +111,6 @@ public class QueryInfo {
      * OrderBy annotation or keyword.
      */
     private static final int[] NONE_STATIC_SORT_ONLY = new int[0];
-
-    /**
-     * Primitive types for numeric values.
-     */
-    static final Set<Class<?>> PRIMITIVE_NUMERIC_TYPES = //
-                    Set.of(long.class, int.class, short.class, byte.class,
-                           double.class, float.class);
-
-    /**
-     * Return types for deleteBy that distinguish delete-only from find-and-delete.
-     */
-    private static final Set<Class<?>> RETURN_TYPES_FOR_DELETE_ONLY = //
-                    Set.of(void.class, Void.class,
-                           boolean.class, Boolean.class,
-                           int.class, Integer.class,
-                           long.class, Long.class,
-                           Number.class);
-
-    /**
-     * Valid types for repository method parameters that specify sort criteria.
-     */
-    static final Set<Class<?>> SORT_PARAM_TYPES = //
-                    Set.of(Order.class, Sort.class, Sort[].class);
-
-    /**
-     * Valid types for repository method parameters after the query parameters.
-     */
-    static final Set<Class<?>> SPECIAL_PARAM_TYPES = //
-                    Set.of(Limit.class, Order.class, PageRequest.class,
-                           Sort.class, Sort[].class);
-
-    /**
-     * Valid types for when a repository method computes an update count
-     */
-    private static final Set<Class<?>> UPDATE_COUNT_TYPES = //
-                    Set.of(boolean.class, Boolean.class,
-                           int.class, Integer.class,
-                           long.class, Long.class,
-                           void.class, Void.class,
-                           Number.class);
-
-    /**
-     * Valid types for jakarta.persistence.Version, except for
-     * java.sql.Timestamp, which is not a valid type in Jakarta Data.
-     */
-    public static final Set<Class<?>> VERSION_TYPES = //
-                    Set.of(Instant.class,
-                           int.class, Integer.class,
-                           LocalDateTime.class,
-                           long.class, Long.class,
-                           short.class, Short.class);
-
-    /**
-     * Mapping of Java primitive class to wrapper class.
-     */
-    private static final Map<Class<?>, Class<?>> WRAPPER_CLASSES = //
-                    Map.of(boolean.class, Boolean.class,
-                           byte.class, Byte.class,
-                           char.class, Character.class,
-                           double.class, Double.class,
-                           float.class, Float.class,
-                           int.class, Integer.class,
-                           long.class, Long.class,
-                           short.class, Short.class,
-                           void.class, Void.class);
-
-    static {
-        for (Entry<Class<?>, Class<?>> e : WRAPPER_CLASSES.entrySet()) {
-            NON_ENTITY_RESULT_TYPES.add(e.getKey()); // primitive classes
-            NON_ENTITY_RESULT_TYPES.add(e.getValue()); // wrapper classes
-        }
-        NON_ENTITY_RESULT_TYPES.add(BigDecimal.class);
-        NON_ENTITY_RESULT_TYPES.add(BigInteger.class);
-        NON_ENTITY_RESULT_TYPES.add(Object.class);
-        NON_ENTITY_RESULT_TYPES.add(String.class);
-    }
 
     /**
      * Information about the type of entity to which the query pertains.
@@ -568,24 +498,6 @@ public class QueryInfo {
     }
 
     /**
-     * Returns true if it is certain the class cannot be an entity
-     * because it is one of the common non-entity result types
-     * or it is an enumeration, interface, or abstract class.
-     * Otherwise, returns false.
-     *
-     * @param c class of result.
-     * @return true if a result of the type might be an entity, otherwise false.
-     */
-    @Trivial
-    public static boolean cannotBeEntity(Class<?> c) {
-        int modifiers;
-        return NON_ENTITY_RESULT_TYPES.contains(c) ||
-               c.isEnum() ||
-               Modifier.isInterface(modifiers = c.getModifiers()) ||
-               Modifier.isAbstract(modifiers);
-    }
-
-    /**
      * Compute the zero-based offset to use as a starting point for a Limit range.
      *
      * @param limit limit that was specified by the application.
@@ -677,7 +589,7 @@ public class QueryInfo {
                                + " to " + toType.getSimpleName());
 
         if (value instanceof Number &&
-            (PRIMITIVE_NUMERIC_TYPES.contains(toType) ||
+            (Util.PRIMITIVE_NUMERIC_TYPES.contains(toType) ||
              Number.class.isAssignableFrom(toType))) {
             // Conversion from one numeric type to another
             try {
@@ -778,7 +690,7 @@ public class QueryInfo {
                 cause = x;
             }
         } else if (value instanceof CharSequence &&
-                   (PRIMITIVE_NUMERIC_TYPES.contains(toType) ||
+                   (Util.PRIMITIVE_NUMERIC_TYPES.contains(toType) ||
                     Number.class.isAssignableFrom(toType))) {
             // Conversion from text to numeric value
             if (int.class.equals(toType) || Integer.class.equals(toType))
@@ -923,6 +835,98 @@ public class QueryInfo {
             list.addAll(results);
         }
         return list;
+    }
+
+    /**
+     * Execute a repository count query.
+     *
+     * @param em   entity manager.
+     * @param args method parameters.
+     * @return the count, converted to a type that is compatible with the
+     *         method signature.
+     * @throws Exception        if an error occurs.
+     * @throws MappingException if the count cannot be converted to the
+     *                              requested type.
+     */
+    @Trivial
+    Object count(EntityManager em, Object... args) throws Exception {
+        final boolean trace = TraceComponent.isAnyTracingEnabled();
+        if (trace && tc.isEntryEnabled())
+            Tr.entry(this, tc, "count", em,
+                     // method args have already been logged if loggable
+                     "to be returned as " + singleType.getName());
+
+        TypedQuery<Long> query = em.createQuery(jpql, Long.class);
+        setParameters(query, args);
+
+        Long count = query.getSingleResult();
+
+        Object returnValue;
+        if (long.class.equals(singleType) ||
+            Long.class.equals(singleType) ||
+            singleType.isAssignableFrom(Long.class)) {
+            returnValue = count;
+        } else {
+            if (int.class.equals(singleType) || Integer.class.equals(singleType))
+                if (count > Integer.MAX_VALUE)
+                    throw exc(MappingException.class,
+                              "CWWKD1048.result.exceeds.max",
+                              count,
+                              method.getName(),
+                              repositoryInterface.getName(),
+                              method.getGenericReturnType().getTypeName(),
+                              "Integer.MAX_VALUE (" + Integer.MAX_VALUE + ')');
+                else
+                    returnValue = count.intValue();
+            else if (short.class.equals(singleType) || Short.class.equals(singleType))
+                if (count > Short.MAX_VALUE)
+                    throw exc(MappingException.class,
+                              "CWWKD1048.result.exceeds.max",
+                              count,
+                              method.getName(),
+                              repositoryInterface.getName(),
+                              method.getGenericReturnType().getTypeName(),
+                              "Short.MAX_VALUE (" + Short.MAX_VALUE + ')');
+                else
+                    returnValue = count.shortValue();
+            else if (byte.class.equals(singleType) || Byte.class.equals(singleType))
+                if (count > Byte.MAX_VALUE)
+                    throw exc(MappingException.class,
+                              "CWWKD1048.result.exceeds.max",
+                              count,
+                              method.getName(),
+                              repositoryInterface.getName(),
+                              method.getGenericReturnType().getTypeName(),
+                              "Byte.MAX_VALUE (" + Byte.MAX_VALUE + ')');
+                else
+                    returnValue = count.byteValue();
+            else if (BigInteger.class.equals(singleType))
+                returnValue = BigInteger.valueOf(count);
+            else if (BigDecimal.class.equals(singleType))
+                returnValue = BigDecimal.valueOf(count);
+            else
+                throw exc(MappingException.class,
+                          "CWWKD1049.count.convert.err",
+                          count,
+                          method.getName(),
+                          repositoryInterface.getName(),
+                          method.getGenericReturnType().getTypeName());
+        }
+
+        Class<?> returnType = method.getReturnType();
+        if (Optional.class.equals(returnType)) {
+            returnValue = Optional.of(returnValue);
+        } else if (CompletableFuture.class.equals(returnType) ||
+                   CompletionStage.class.equals(returnType)) {
+            returnValue = CompletableFuture.completedFuture(returnValue);
+        }
+
+        if (trace && tc.isEntryEnabled())
+            if (count == returnValue)
+                Tr.exit(this, tc, "count", returnValue);
+            else
+                Tr.exit(this, tc, "count", count + " converted to " + returnValue);
+        return returnValue;
     }
 
     /**
@@ -1084,6 +1088,24 @@ public class QueryInfo {
     }
 
     /**
+     * Create a new NonUniqueResultException.
+     *
+     * @param numResults number of results.
+     * @return the NonUniqueResultException.
+     */
+    @Trivial
+    NonUniqueResultException excNonUniqueResult(int numResults) {
+        throw exc(NonUniqueResultException.class,
+                  method.getName(),
+                  repositoryInterface.getName(),
+                  method.getGenericReturnType().getTypeName(),
+                  numResults,
+                  List.of(// In a future release: @Find @First findByX(...)
+                          "findFirstByX(...)",
+                          "findByX(..., Limit.of(1))"));
+    }
+
+    /**
      * Constructs the UnsupportedOperationException for the general error where
      * a repository method is unrecognized and log the error.
      *
@@ -1100,6 +1122,186 @@ public class QueryInfo {
                    List.of("Connection", "DataSource", "EntityManager"),
                    List.of("count", "delete", "exists", "find"),
                    entityInfo.getExampleMethodNames());
+    }
+
+    /**
+     * Finds and updates entities (or records) in the database.
+     *
+     * @param arg the entity or record, or array/Iterable/Stream of entity or record
+     * @param em  the entity manager.
+     * @return the updated entities, using the return type that is required by the
+     *         repository Update method signature.
+     * @throws OptimisticLockingFailureException if an entity is not found in the database.
+     * @throws Exception                         if an error occurs.
+     */
+    Object findAndUpdate(Object arg, EntityManager em) throws Exception {
+        List<Object> results;
+
+        boolean hasSingularEntityParam = false;
+        if (entityParamType.isArray()) {
+            int length = Array.getLength(arg);
+            results = new ArrayList<>(length);
+            for (int i = 0; i < length; i++)
+                results.add(findAndUpdateOne(Array.get(arg, i), em));
+        } else {
+            arg = arg instanceof Stream //
+                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
+                            : arg;
+
+            results = new ArrayList<>();
+            if (arg instanceof Iterable) {
+                for (Object e : ((Iterable<?>) arg))
+                    results.add(findAndUpdateOne(e, em));
+            } else {
+                hasSingularEntityParam = true;
+                results = new ArrayList<>(1);
+                results.add(findAndUpdateOne(arg, em));
+            }
+        }
+        em.flush();
+
+        Class<?> returnType = method.getReturnType();
+        if (void.class.equals(returnType) || Void.class.equals(returnType))
+            return null;
+
+        if (entityInfo.recordClass != null)
+            for (int i = 0; i < results.size(); i++)
+                results.set(i, entityInfo.toRecord(results.get(i)));
+
+        Object returnValue;
+        if (returnArrayType != null) {
+            Object[] newArray = (Object[]) Array.newInstance(returnArrayType, results.size());
+            returnValue = results.toArray(newArray);
+        } else {
+            if (multiType == null)
+                if (results.size() == 1)
+                    returnValue = results.get(0);
+                else if (results.isEmpty())
+                    returnValue = null;
+                else
+                    throw excNonUniqueResult(results.size());
+            else if (multiType.isInstance(results))
+                returnValue = results;
+            else if (Stream.class.equals(multiType))
+                returnValue = results.stream();
+            else if (Iterable.class.isAssignableFrom(multiType))
+                returnValue = convertToIterable(results, null, multiType);
+            else if (Iterator.class.equals(multiType))
+                returnValue = results.iterator();
+            else
+                throw exc(MappingException.class,
+                          "CWWKD1003.lifecycle.rtrn.err",
+                          method.getGenericReturnType().getTypeName(),
+                          method.getName(),
+                          repositoryInterface.getName(),
+                          "Update",
+                          lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                               hasSingularEntityParam,
+                                               false));
+        }
+
+        if (Optional.class.equals(returnType)) {
+            returnValue = returnValue == null ? Optional.empty() : Optional.of(returnValue);
+        } else if (CompletableFuture.class.equals(returnType) || CompletionStage.class.equals(returnType)) {
+            returnValue = CompletableFuture.completedFuture(returnValue); // useful for @Asynchronous
+        } else if (returnValue != null && !returnType.isInstance(returnValue)) {
+            throw exc(MappingException.class,
+                      "CWWKD1003.lifecycle.rtrn.err",
+                      method.getGenericReturnType().getTypeName(),
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      "Update",
+                      lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                           hasSingularEntityParam,
+                                           false));
+        }
+
+        return returnValue;
+    }
+
+    /**
+     * Finds an entity (or record) in the database, locks it for subsequent update,
+     * and updates the entity found in the database to match the desired state
+     * of the supplied entity.
+     *
+     * @param e  the entity or record.
+     * @param em the entity manager.
+     * @return the entity that is written to the database. Never null.
+     * @throws OptimisticLockingException if the entity is not found.
+     * @throws Exception                  if an error occurs.
+     */
+    private Object findAndUpdateOne(Object e, EntityManager em) throws Exception {
+        String jpql = this.jpql;
+
+        int versionParamIndex = entityInfo.idClassAttributeAccessors == null //
+                        ? 2 //
+                        : (entityInfo.idClassAttributeAccessors.size() + 1);
+        Object version = null;
+        if (entityInfo.versionAttributeName != null) {
+            version = getAttribute(e, entityInfo.versionAttributeName);
+            if (version == null)
+                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
+        }
+
+        Object id = null;
+        String idAttributeName = null;
+        if (entityInfo.idClassAttributeAccessors == null) {
+            idAttributeName = entityInfo.attributeNames.get(ID);
+            id = getAttribute(e, idAttributeName);
+            if (id == null) {
+                jpql = jpql.replace("=?" + (versionParamIndex - 1),
+                                    " IS NULL");
+                if (version != null)
+                    jpql = jpql.replace("=?" + versionParamIndex,
+                                        "=?" + (versionParamIndex - 1));
+            }
+        }
+
+        if (TraceComponent.isAnyTracingEnabled() && jpql != this.jpql)
+            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
+
+        Class<?> entityClass = singleType.equals(entityInfo.recordClass) //
+                        ? entityInfo.entityClass //
+                        : singleType;
+
+        TypedQuery<?> query = em.createQuery(jpql, entityClass);
+        query.setLockMode(LockModeType.PESSIMISTIC_WRITE);
+
+        if (entityInfo.idClassAttributeAccessors == null) {
+            int p = 1;
+            if (id != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(id));
+                query.setParameter(p++, id);
+            }
+            if (version != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(version));
+                query.setParameter(p, version);
+            }
+        } else {
+            setParametersFromIdClassAndVersion(1, query, e, version);
+        }
+
+        List<?> results = query.getResultList();
+
+        if (results.isEmpty()) {
+            List<String> entityProps = new ArrayList<>(2);
+            if (id != null)
+                entityProps.add(loggableAppend(idAttributeName, "=", id));
+            if (entityInfo.versionAttributeName != null && version != null)
+                entityProps.add(loggableAppend(entityInfo.versionAttributeName,
+                                               "=", version));
+            throw exc(OptimisticLockingFailureException.class,
+                      "CWWKD1050.opt.lock.exc",
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      e.getClass().getName(),
+                      entityProps,
+                      RepositoryImpl.LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES);
+        }
+
+        return em.merge(toEntity(e));
     }
 
     /**
@@ -1845,10 +2047,10 @@ public class QueryInfo {
         Class<?> singleType = this.singleType;
 
         if (singleType.isPrimitive())
-            singleType = wrapperClassIfPrimitive(singleType);
+            singleType = Util.wrapperClassIfPrimitive(singleType);
 
         if (type == Type.FIND_AND_DELETE &&
-            !(singleType.isAssignableFrom(wrapperClassIfPrimitive(entityInfo.idType)) ||
+            !(singleType.isAssignableFrom(Util.wrapperClassIfPrimitive(entityInfo.idType)) ||
               singleType.isAssignableFrom(entityInfo.getType()))) {
             throw exc(MappingException.class,
                       "CWWKD1006.delete.rtrn.err",
@@ -1871,7 +2073,7 @@ public class QueryInfo {
                 for (Map.Entry<String, Class<?>> entry : entityInfo.attributeTypes.entrySet()) {
                     Class<?> attributeType = entry.getValue();
                     if (attributeType.isPrimitive())
-                        attributeType = wrapperClassIfPrimitive(attributeType);
+                        attributeType = Util.wrapperClassIfPrimitive(attributeType);
                     if (singleType.isAssignableFrom(attributeType))
                         if (singleAttributeName == null)
                             singleAttributeName = entry.getKey();
@@ -2097,7 +2299,7 @@ public class QueryInfo {
         StringBuilder q;
 
         if (entityInfo.attributeNamesForEntityUpdate != null &&
-            UPDATE_COUNT_TYPES.contains(singleType)) {
+            Util.UPDATE_COUNT_TYPES.contains(singleType)) {
             setType(Update.class, Type.UPDATE_WITH_ENTITY_PARAM);
 
             q = new StringBuilder(100) //
@@ -3255,6 +3457,127 @@ public class QueryInfo {
     }
 
     /**
+     * Inserts entities (or records) into the database.
+     * An error is raised if any of the entities (or records) already exist
+     * in the database.
+     *
+     * @param arg the entity or record, or array/Iterable/Stream of entity or record
+     * @param em  the entity manager
+     * @return the updated entities, using the return type that is required by the
+     *         Insert method signature.
+     * @throws Exception if an error occurs.
+     */
+    Object insert(Object arg, EntityManager em) throws Exception {
+        boolean resultVoid = void.class.equals(singleType) ||
+                             Void.class.equals(singleType);
+        ArrayList<Object> results;
+
+        boolean hasSingularEntityParam = false;
+        if (entityParamType.isArray()) {
+            int length = Array.getLength(arg);
+            results = resultVoid ? null : new ArrayList<>(length);
+            for (int i = 0; i < length; i++) {
+                Object entity = toEntity(Array.get(arg, i));
+                em.persist(entity);
+                if (results != null)
+                    results.add(entity);
+            }
+            em.flush();
+        } else {
+            arg = arg instanceof Stream //
+                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
+                            : arg;
+
+            if (arg instanceof Iterable) {
+                results = resultVoid ? null : new ArrayList<>();
+                for (Object e : ((Iterable<?>) arg)) {
+                    Object entity = toEntity(e);
+                    em.persist(entity);
+                    if (results != null)
+                        results.add(entity);
+                }
+                em.flush();
+            } else {
+                hasSingularEntityParam = true;
+                results = resultVoid ? null : new ArrayList<>(1);
+                Object entity = toEntity(arg);
+                em.persist(entity);
+                em.flush();
+                if (results != null)
+                    results.add(entity);
+            }
+        }
+
+        Class<?> returnType = method.getReturnType();
+        Object returnValue;
+        if (resultVoid) {
+            returnValue = null;
+        } else {
+            if (entityInfo.recordClass != null)
+                for (int i = 0; i < results.size(); i++)
+                    results.set(i, entityInfo.toRecord(results.get(i)));
+
+            if (returnArrayType != null) {
+                Object[] newArray = (Object[]) Array.newInstance(returnArrayType,
+                                                                 results.size());
+                returnValue = results.toArray(newArray);
+            } else {
+                if (multiType == null)
+                    if (results.size() == 1)
+                        returnValue = results.get(0);
+                    else if (results.isEmpty())
+                        returnValue = null;
+                    else
+                        // TODO temporarily reusing this message, which is somewhat close.
+                        // Need a more specific one, possibly resuing parts of CWWKD1054,
+                        // but without the Find suggestions
+                        throw exc(ClassCastException.class,
+                                  "CWWKD1046.result.convert.err",
+                                  entityInfo.getType().getSimpleName() + "[]",
+                                  method.getName(),
+                                  repositoryInterface.getName(),
+                                  method.getGenericReturnType().getTypeName());
+                else if (multiType.isInstance(results))
+                    returnValue = results;
+                else if (Stream.class.equals(multiType))
+                    returnValue = results.stream();
+                else if (Iterable.class.isAssignableFrom(multiType))
+                    returnValue = convertToIterable(results, null, multiType);
+                else if (Iterator.class.equals(multiType))
+                    returnValue = results.iterator();
+                else
+                    throw exc(MappingException.class,
+                              "CWWKD1003.lifecycle.rtrn.err",
+                              method.getGenericReturnType().getTypeName(),
+                              method.getName(),
+                              repositoryInterface.getName(),
+                              "Insert",
+                              lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                                   hasSingularEntityParam,
+                                                   false));
+            }
+        }
+
+        if (CompletableFuture.class.equals(returnType) ||
+            CompletionStage.class.equals(returnType)) {
+            // useful for @Asynchronous
+            returnValue = CompletableFuture.completedFuture(returnValue);
+        } else if (!resultVoid && !returnType.isInstance(returnValue)) {
+            throw exc(MappingException.class,
+                      "CWWKD1003.lifecycle.rtrn.err",
+                      method.getGenericReturnType().getTypeName(),
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      "Insert",
+                      lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                           hasSingularEntityParam,
+                                           false));
+        }
+
+        return returnValue;
+    }
+
+    /**
      * Determines whether a delete operation is find-and-delete (true) or delete only (false).
      * The determination is made based on the return type, with multiple and Optional results
      * indicating find-and-delete, and void or singular results that are boolean or a numeric
@@ -3272,7 +3595,7 @@ public class QueryInfo {
 
         boolean isFindAndDelete = isOptional
                                   || multiType != null
-                                  || !RETURN_TYPES_FOR_DELETE_ONLY.contains(singleType);
+                                  || !Util.RETURN_TYPES_FOR_DELETE_ONLY.contains(singleType);
 
         if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
             Tr.debug(this, tc, "isFindAndDelete? " + isFindAndDelete +
@@ -3285,8 +3608,8 @@ public class QueryInfo {
                 && !type.equals(entityInfo.entityClass)
                 && !type.equals(entityInfo.recordClass)
                 && !type.equals(Object.class)
-                && !wrapperClassIfPrimitive(singleType) //
-                                .equals(wrapperClassIfPrimitive(entityInfo.idType))) {
+                && !Util.wrapperClassIfPrimitive(singleType) //
+                                .equals(Util.wrapperClassIfPrimitive(entityInfo.idType))) {
                 throw exc(MappingException.class,
                           "CWWKD1006.delete.rtrn.err",
                           method.getGenericReturnType().getTypeName(),
@@ -3470,6 +3793,107 @@ public class QueryInfo {
     }
 
     /**
+     * Removes the entity (or record) from the database if its attributes
+     * match the database.
+     *
+     * @param e  the entity or record.
+     * @param em the entity manager.
+     * @return the number of entities deleted (1 or 0).
+     * @throws Exception if an error occurs or the repository method return type is
+     *                       void and the entity (or correct version of the entity)
+     *                       was not found.
+     */
+    int remove(Object e, EntityManager em) throws Exception {
+        Class<?> entityClass = entityInfo.getType();
+
+        if (e == null)
+            throw exc(NullPointerException.class,
+                      "CWWKD1015.null.entity.param",
+                      method.getName(),
+                      repositoryInterface.getName());
+
+        if (!entityClass.isInstance(e))
+            throw exc(IllegalArgumentException.class,
+                      "CWWKD1016.incompat.entity.param",
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      entityClass.getName(),
+                      e.getClass().getName());
+
+        String jpql = this.jpql;
+
+        int versionParamIndex = (entityInfo.idClassAttributeAccessors == null ? 1 : entityInfo.idClassAttributeAccessors.size()) + 1;
+        Object version = null;
+        if (entityInfo.versionAttributeName != null) {
+            version = getAttribute(e, entityInfo.versionAttributeName);
+            if (version == null)
+                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
+        }
+
+        Object id = null;
+        String idAttributeName = null;
+        if (entityInfo.idClassAttributeAccessors == null) {
+            idAttributeName = getAttributeName(ID, true);
+            id = getAttribute(e, idAttributeName);
+            if (id == null) {
+                jpql = jpql.replace("=?" + (versionParamIndex - 1), " IS NULL");
+                if (version != null)
+                    jpql = jpql.replace("=?" + versionParamIndex, "=?" + (versionParamIndex - 1));
+            }
+        }
+
+        if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled()
+            && jpql != this.jpql)
+            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
+
+        TypedQuery<?> delete = em.createQuery(jpql, entityInfo.entityClass);
+
+        if (entityInfo.idClassAttributeAccessors == null) {
+            int p = 1;
+            if (id != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(id));
+                delete.setParameter(p++, id);
+            }
+            if (version != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(version));
+                delete.setParameter(p, version);
+            }
+        } else {
+            setParametersFromIdClassAndVersion(1, delete, e, version);
+        }
+
+        int numDeleted = delete.executeUpdate();
+
+        if (numDeleted == 0) {
+            Class<?> returnType = method.getReturnType();
+            if (void.class.equals(returnType) || Void.class.equals(returnType)) {
+                if (idAttributeName == null)
+                    idAttributeName = ID;
+                List<String> entityProps = new ArrayList<>(2);
+                if (id != null)
+                    entityProps.add(loggableAppend(idAttributeName,
+                                                   "=", id));
+                if (entityInfo.versionAttributeName != null && version != null)
+                    entityProps.add(loggableAppend(entityInfo.versionAttributeName,
+                                                   "=", version));
+                throw exc(OptimisticLockingFailureException.class,
+                          "CWWKD1050.opt.lock.exc",
+                          method.getName(),
+                          repositoryInterface.getName(),
+                          e.getClass().getName(),
+                          entityProps,
+                          RepositoryImpl.LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES);
+            }
+        } else if (numDeleted > 1) {
+            throw new DataException("Found " + numDeleted + " matching entities."); // ought to be unreachable
+        }
+
+        return numDeleted;
+    }
+
+    /**
      * Pagination is only possible if results are ordered.
      *
      * If the repository method has parameters to supply an order and these are
@@ -3516,6 +3940,116 @@ public class QueryInfo {
                       method.getName(),
                       repositoryInterface.getName(),
                       method.getGenericReturnType().getTypeName());
+    }
+
+    /**
+     * Saves entities (or records) to the database, which can involve an update
+     * or an insert, depending on whether the entity already exists.
+     *
+     * @param arg the entity or record, or array/Iterable/Stream of entity or record
+     * @param em  the entity manager.
+     * @return the updated entities, using the return type that is required by the
+     *         repository Save method signature.
+     * @throws Exception if an error occurs.
+     */
+    Object save(Object arg, EntityManager em) throws Exception {
+        boolean resultVoid = void.class.equals(singleType) ||
+                             Void.class.equals(singleType);
+        List<Object> results;
+
+        boolean hasSingularEntityParam = false;
+        if (entityParamType.isArray()) {
+            results = new ArrayList<>();
+            int length = Array.getLength(arg);
+            for (int i = 0; i < length; i++)
+                results.add(em.merge(toEntity(Array.get(arg, i))));
+            em.flush();
+        } else {
+            arg = arg instanceof Stream //
+                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
+                            : arg;
+
+            if (Iterable.class.isAssignableFrom(entityParamType)) {
+                results = new ArrayList<>();
+                for (Object e : ((Iterable<?>) arg))
+                    results.add(em.merge(toEntity(e)));
+                em.flush();
+            } else {
+                hasSingularEntityParam = true;
+                results = resultVoid ? null : new ArrayList<>(1);
+                Object entity = em.merge(toEntity(arg));
+                if (results != null)
+                    results.add(entity);
+                em.flush();
+            }
+        }
+
+        Class<?> returnType = method.getReturnType();
+        Object returnValue;
+        if (resultVoid) {
+            returnValue = null;
+        } else {
+            if (entityInfo.recordClass != null)
+                for (int i = 0; i < results.size(); i++)
+                    results.set(i, entityInfo.toRecord(results.get(i)));
+
+            if (returnArrayType != null) {
+                Object[] newArray = (Object[]) Array.newInstance(returnArrayType, results.size());
+                returnValue = results.toArray(newArray);
+            } else {
+                if (multiType == null)
+                    if (results.size() == 1)
+                        returnValue = results.get(0);
+                    else if (results.isEmpty())
+                        returnValue = null;
+                    else
+                        // TODO temporarily reusing this message, which is somewhat close.
+                        // Need a more specific one, possibly resuing parts of CWWKD1054,
+                        // but without the Find suggestions
+                        throw exc(ClassCastException.class,
+                                  "CWWKD1046.result.convert.err",
+                                  entityInfo.getType().getSimpleName() + "[]",
+                                  method.getName(),
+                                  repositoryInterface.getName(),
+                                  method.getGenericReturnType().getTypeName());
+                else if (multiType.isInstance(results))
+                    returnValue = results;
+                else if (Stream.class.equals(multiType))
+                    returnValue = results.stream();
+                else if (Iterable.class.isAssignableFrom(multiType))
+                    returnValue = convertToIterable(results, null, multiType);
+                else if (Iterator.class.equals(multiType))
+                    returnValue = results.iterator();
+                else
+                    throw exc(MappingException.class,
+                              "CWWKD1003.lifecycle.rtrn.err",
+                              method.getGenericReturnType().getTypeName(),
+                              method.getName(),
+                              repositoryInterface.getName(),
+                              "Save",
+                              lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                                   hasSingularEntityParam,
+                                                   false));
+            }
+        }
+
+        if (CompletableFuture.class.equals(returnType) ||
+            CompletionStage.class.equals(returnType)) {
+            // useful for @Asynchronous
+            returnValue = CompletableFuture.completedFuture(returnValue);
+        } else if (!resultVoid && !returnType.isInstance(returnValue)) {
+            throw exc(MappingException.class,
+                      "CWWKD1003.lifecycle.rtrn.err",
+                      method.getGenericReturnType().getTypeName(),
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      "Save",
+                      lifeCycleReturnTypes(results.get(0).getClass().getSimpleName(),
+                                           hasSingularEntityParam,
+                                           false));
+        }
+
+        return returnValue;
     }
 
     /**
@@ -3964,6 +4498,91 @@ public class QueryInfo {
     }
 
     /**
+     * Updates the entity (or record) from the database if its attributes match the database.
+     *
+     * @param e  the entity or record.
+     * @param em the entity manager.
+     * @return the number of entities updated (1 or 0).
+     * @throws Exception if an error occurs.
+     */
+    int update(Object e, EntityManager em) throws Exception {
+        Class<?> entityClass = entityInfo.getType();
+
+        if (e == null)
+            throw exc(NullPointerException.class,
+                      "CWWKD1015.null.entity.param",
+                      method.getName(),
+                      repositoryInterface.getName());
+
+        if (!entityClass.isInstance(e))
+            throw exc(IllegalArgumentException.class,
+                      "CWWKD1016.incompat.entity.param",
+                      method.getName(),
+                      repositoryInterface.getName(),
+                      entityClass.getName(),
+                      e.getClass().getName());
+
+        String jpql = this.jpql;
+        Set<String> attrsToUpdate = entityInfo.attributeNamesForEntityUpdate;
+
+        int versionParamIndex = entityInfo.idClassAttributeAccessors == null //
+                        ? (attrsToUpdate.size() + 2) //
+                        : (attrsToUpdate.size() +
+                           entityInfo.idClassAttributeAccessors.size() + 1);
+        Object version = null;
+        if (entityInfo.versionAttributeName != null) {
+            version = getAttribute(e, entityInfo.versionAttributeName);
+            if (version == null)
+                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
+        }
+
+        Object id = null;
+        String idAttributeName = null;
+        if (entityInfo.idClassAttributeAccessors == null) {
+            idAttributeName = entityInfo.attributeNames.get(ID);
+            id = getAttribute(e, idAttributeName);
+            if (id == null) {
+                jpql = jpql.replace("=?" + (versionParamIndex - 1), " IS NULL");
+                if (version != null)
+                    jpql = jpql.replace("=?" + versionParamIndex, "=?" + (versionParamIndex - 1));
+            }
+        }
+
+        if (TraceComponent.isAnyTracingEnabled() && jpql != this.jpql)
+            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
+
+        TypedQuery<?> update = em.createQuery(jpql, entityInfo.entityClass);
+
+        // parameters for entity attributes to update:
+        int p = 1;
+        for (String attrName : attrsToUpdate)
+            setParameter(p++, update, e, attrName);
+
+        // id and version parameters
+        if (entityInfo.idClassAttributeAccessors == null) {
+            if (id != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(id));
+                update.setParameter(p++, id);
+            }
+
+            if (entityInfo.versionAttributeName != null && version != null) {
+                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
+                    Tr.debug(tc, "set ?" + p + ' ' + loggable(version));
+                update.setParameter(p++, version);
+            }
+        } else { // has IdClass
+            setParametersFromIdClassAndVersion(p, update, e, version);
+        }
+
+        int numUpdated = update.executeUpdate();
+
+        if (numUpdated > 1) // ought to be unreachable
+            throw new DataException("Found " + numUpdated + " matching entities.");
+        return numUpdated;
+    }
+
+    /**
      * Ensure that the annotations are valid together on the same repository method.
      *
      * @param delete  The Delete annotation if present, otherwise null.
@@ -4084,17 +4703,5 @@ public class QueryInfo {
         q.validateParams = validateParams;
         q.validateParams = validateResult;
         return q;
-    }
-
-    /**
-     * Returns the wrapper class if a primitive class, otherwise the same class.
-     *
-     * @param c class that is possibly a primitive class.
-     * @return wrapper class for a primitive, otherwise the same class that was supplied as a parameter.
-     */
-    @Trivial
-    private static final Class<?> wrapperClassIfPrimitive(Class<?> c) {
-        Class<?> w = WRAPPER_CLASSES.get(c);
-        return w == null ? c : w;
     }
 }
