+++ dev/io.openliberty.data.internal.persistence/src/io/openliberty/data/internal/persistence/RepositoryImpl.java
@@ -21,8 +21,6 @@ import java.lang.reflect.InvocationHandler;
 import java.lang.reflect.Member;
 import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.SQLIntegrityConstraintViolationException;
@@ -40,7 +38,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Optional;
-import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.CompletionStage;
@@ -97,7 +94,7 @@ public class RepositoryImpl<R> implements InvocationHandler {
 
     private static final ThreadLocal<Deque<AutoCloseable>> defaultMethodResources = new ThreadLocal<>();
 
-    private static final List<String> LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES = //
+    static final List<String> LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES = //
                     List.of(Insert.class.getSimpleName(),
                             Save.class.getSimpleName(),
                             Update.class.getSimpleName());
@@ -276,102 +273,6 @@ public class RepositoryImpl<R> implements InvocationHandler {
     }
 
     /**
-     * Execute a repository count query.
-     *
-     * @param em        entity manager.
-     * @param queryInfo query information.
-     * @param args      method parameters.
-     * @return the count, converted to a type that is compatible with the
-     *         method signature.
-     * @throws Exception        if an error occurs.
-     * @throws MappingException if the count cannot be converted to the
-     *                              requested type.
-     */
-    @Trivial
-    private Object count(EntityManager em,
-                         QueryInfo queryInfo,
-                         Object... args) throws Exception {
-        Class<?> type = queryInfo.singleType;
-
-        final boolean trace = TraceComponent.isAnyTracingEnabled();
-        if (trace && tc.isEntryEnabled())
-            Tr.entry(this, tc, "count", em, queryInfo,
-                     // method args have already been logged if loggable
-                     "to be returned as " + type.getName());
-
-        TypedQuery<Long> query = em.createQuery(queryInfo.jpql, Long.class);
-        queryInfo.setParameters(query, args);
-
-        Long count = query.getSingleResult();
-
-        Object returnValue;
-        if (long.class.equals(type) ||
-            Long.class.equals(type) ||
-            type.isAssignableFrom(Long.class)) {
-            returnValue = count;
-        } else {
-            if (int.class.equals(type) || Integer.class.equals(type))
-                if (count > Integer.MAX_VALUE)
-                    throw exc(MappingException.class,
-                              "CWWKD1048.result.exceeds.max",
-                              count,
-                              queryInfo.method.getName(),
-                              repositoryInterface.getName(),
-                              queryInfo.method.getGenericReturnType().getTypeName(),
-                              "Integer.MAX_VALUE (" + Integer.MAX_VALUE + ')');
-                else
-                    returnValue = count.intValue();
-            else if (short.class.equals(type) || Short.class.equals(type))
-                if (count > Short.MAX_VALUE)
-                    throw exc(MappingException.class,
-                              "CWWKD1048.result.exceeds.max",
-                              count,
-                              queryInfo.method.getName(),
-                              repositoryInterface.getName(),
-                              queryInfo.method.getGenericReturnType().getTypeName(),
-                              "Short.MAX_VALUE (" + Short.MAX_VALUE + ')');
-                else
-                    returnValue = count.shortValue();
-            else if (byte.class.equals(type) || Byte.class.equals(type))
-                if (count > Byte.MAX_VALUE)
-                    throw exc(MappingException.class,
-                              "CWWKD1048.result.exceeds.max",
-                              count,
-                              queryInfo.method.getName(),
-                              repositoryInterface.getName(),
-                              queryInfo.method.getGenericReturnType().getTypeName(),
-                              "Byte.MAX_VALUE (" + Byte.MAX_VALUE + ')');
-                else
-                    returnValue = count.byteValue();
-            else if (BigInteger.class.equals(type))
-                returnValue = BigInteger.valueOf(count);
-            else if (BigDecimal.class.equals(type))
-                returnValue = BigDecimal.valueOf(count);
-            else
-                throw exc(MappingException.class,
-                          "CWWKD1049.count.convert.err",
-                          count,
-                          queryInfo.method.getName(),
-                          repositoryInterface.getName(),
-                          queryInfo.method.getGenericReturnType().getTypeName());
-        }
-
-        Class<?> returnType = queryInfo.method.getReturnType();
-        if (Optional.class.equals(returnType)) {
-            returnValue = Optional.of(returnValue);
-        } else if (CompletableFuture.class.equals(returnType) || CompletionStage.class.equals(returnType)) {
-            returnValue = CompletableFuture.completedFuture(returnValue);
-        }
-
-        if (trace && tc.isEntryEnabled())
-            if (count == returnValue)
-                Tr.exit(this, tc, "count", returnValue);
-            else
-                Tr.exit(this, tc, "count", count + " converted to " + returnValue);
-        return returnValue;
-    }
-
-    /**
      * Create a new EmptyResultException.
      *
      * @param method repository method that unexpectedly finds an empty result.
@@ -389,25 +290,6 @@ public class RepositoryImpl<R> implements InvocationHandler {
                            Page.class.getSimpleName(),
                            CursoredPage.class.getSimpleName(),
                            Stream.class.getSimpleName()));
-    }
-
-    /**
-     * Create a new NonUniqueResultException.
-     *
-     * @param method repository method that unexpectedly finds multiple results.
-     * @return the NonUniqueResultException.
-     */
-    @Trivial
-    private NonUniqueResultException excNonUniqueResult(QueryInfo queryInfo,
-                                                        int numResults) {
-        throw exc(NonUniqueResultException.class,
-                  queryInfo.method.getName(),
-                  repositoryInterface.getName(),
-                  queryInfo.method.getGenericReturnType().getTypeName(),
-                  numResults,
-                  List.of(// In a future release: @Find @First findByX(...)
-                          "findFirstByX(...)",
-                          "findByX(..., Limit.of(1))"));
     }
 
     /**
@@ -496,187 +378,6 @@ public class RepositoryImpl<R> implements InvocationHandler {
     }
 
     /**
-     * Finds and updates entities (or records) in the database.
-     *
-     * @param arg       the entity or record, or array or Iterable or Stream of entity or record.
-     * @param queryInfo query information.
-     * @param em        the entity manager.
-     * @return the updated entities, using the return type that is required by the repository Update method signature.
-     * @throws OptimisticLockingFailureException if an entity is not found in the database.
-     * @throws Exception                         if an error occurs.
-     */
-    private Object findAndUpdate(Object arg, QueryInfo queryInfo, EntityManager em) throws Exception {
-        List<Object> results;
-
-        boolean hasSingularEntityParam = false;
-        if (queryInfo.entityParamType.isArray()) {
-            int length = Array.getLength(arg);
-            results = new ArrayList<>(length);
-            for (int i = 0; i < length; i++)
-                results.add(findAndUpdateOne(Array.get(arg, i), queryInfo, em));
-        } else {
-            arg = arg instanceof Stream //
-                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
-                            : arg;
-
-            results = new ArrayList<>();
-            if (arg instanceof Iterable) {
-                for (Object e : ((Iterable<?>) arg))
-                    results.add(findAndUpdateOne(e, queryInfo, em));
-            } else {
-                hasSingularEntityParam = true;
-                results = new ArrayList<>(1);
-                results.add(findAndUpdateOne(arg, queryInfo, em));
-            }
-        }
-        em.flush();
-
-        Class<?> returnType = queryInfo.method.getReturnType();
-        if (void.class.equals(returnType) || Void.class.equals(returnType))
-            return null;
-
-        if (queryInfo.entityInfo.recordClass != null)
-            for (int i = 0; i < results.size(); i++)
-                results.set(i, queryInfo.entityInfo.toRecord(results.get(i)));
-
-        Object returnValue;
-        if (queryInfo.returnArrayType != null) {
-            Object[] newArray = (Object[]) Array.newInstance(queryInfo.returnArrayType, results.size());
-            returnValue = results.toArray(newArray);
-        } else {
-            Class<?> multiType = queryInfo.multiType;
-            if (multiType == null)
-                if (results.size() == 1)
-                    returnValue = results.get(0);
-                else if (results.isEmpty())
-                    returnValue = null;
-                else
-                    throw excNonUniqueResult(queryInfo, results.size());
-            else if (multiType.isInstance(results))
-                returnValue = results;
-            else if (Stream.class.equals(multiType))
-                returnValue = results.stream();
-            else if (Iterable.class.isAssignableFrom(multiType))
-                returnValue = queryInfo.convertToIterable(results, null, multiType);
-            else if (Iterator.class.equals(multiType))
-                returnValue = results.iterator();
-            else
-                throw exc(MappingException.class,
-                          "CWWKD1003.lifecycle.rtrn.err",
-                          queryInfo.method.getGenericReturnType().getTypeName(),
-                          queryInfo.method.getName(),
-                          repositoryInterface.getName(),
-                          "Update",
-                          getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                              hasSingularEntityParam,
-                                              false));
-        }
-
-        if (Optional.class.equals(returnType)) {
-            returnValue = returnValue == null ? Optional.empty() : Optional.of(returnValue);
-        } else if (CompletableFuture.class.equals(returnType) || CompletionStage.class.equals(returnType)) {
-            returnValue = CompletableFuture.completedFuture(returnValue); // useful for @Asynchronous
-        } else if (returnValue != null && !returnType.isInstance(returnValue)) {
-            throw exc(MappingException.class,
-                      "CWWKD1003.lifecycle.rtrn.err",
-                      queryInfo.method.getGenericReturnType().getTypeName(),
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      "Update",
-                      getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                          hasSingularEntityParam,
-                                          false));
-        }
-
-        return returnValue;
-    }
-
-    /**
-     * Finds an entity (or record) in the database, locks it for subsequent update,
-     * and updates the entity found in the database to match the desired state of the supplied entity.
-     *
-     * @param e         the entity or record.
-     * @param queryInfo query information.
-     * @param em        the entity manager.
-     * @return the entity that is written to the database. Never null.
-     * @throws OptimisticLockingException if the entity is not found.
-     * @throws Exception                  if an error occurs.
-     */
-    private Object findAndUpdateOne(Object e, QueryInfo queryInfo, EntityManager em) throws Exception {
-        String jpql = queryInfo.jpql;
-        EntityInfo entityInfo = queryInfo.entityInfo;
-
-        int versionParamIndex = entityInfo.idClassAttributeAccessors == null //
-                        ? 2 //
-                        : (entityInfo.idClassAttributeAccessors.size() + 1);
-        Object version = null;
-        if (entityInfo.versionAttributeName != null) {
-            version = queryInfo.getAttribute(e, entityInfo.versionAttributeName);
-            if (version == null)
-                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
-        }
-
-        Object id = null;
-        String idAttributeName = null;
-        if (entityInfo.idClassAttributeAccessors == null) {
-            idAttributeName = entityInfo.attributeNames.get(ID);
-            id = queryInfo.getAttribute(e, idAttributeName);
-            if (id == null) {
-                jpql = jpql.replace("=?" + (versionParamIndex - 1), " IS NULL");
-                if (version != null)
-                    jpql = jpql.replace("=?" + versionParamIndex, "=?" + (versionParamIndex - 1));
-            }
-        }
-
-        if (TraceComponent.isAnyTracingEnabled() && jpql != queryInfo.jpql)
-            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
-
-        Class<?> entityClass = queryInfo.singleType.equals(entityInfo.recordClass) //
-                        ? entityInfo.entityClass //
-                        : queryInfo.singleType;
-
-        TypedQuery<?> query = em.createQuery(jpql, entityClass);
-        query.setLockMode(LockModeType.PESSIMISTIC_WRITE);
-
-        if (entityInfo.idClassAttributeAccessors == null) {
-            int p = 1;
-            if (id != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(id));
-                query.setParameter(p++, id);
-            }
-            if (version != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(version));
-                query.setParameter(p, version);
-            }
-        } else {
-            queryInfo.setParametersFromIdClassAndVersion(1, query, e, version);
-        }
-
-        List<?> results = query.getResultList();
-
-        if (results.isEmpty()) {
-            List<String> entityProps = new ArrayList<>(2);
-            if (id != null)
-                entityProps.add(queryInfo.loggableAppend(idAttributeName,
-                                                         "=", id));
-            if (entityInfo.versionAttributeName != null && version != null)
-                entityProps.add(queryInfo.loggableAppend(entityInfo.versionAttributeName,
-                                                         "=", version));
-            throw exc(OptimisticLockingFailureException.class,
-                      "CWWKD1050.opt.lock.exc",
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      e.getClass().getName(),
-                      entityProps,
-                      LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES);
-        }
-
-        return em.merge(queryInfo.toEntity(e));
-    }
-
-    /**
      * Return a name for the parameter, suitable for display in an NLS message.
      *
      * @param param parameter
@@ -751,144 +452,6 @@ public class RepositoryImpl<R> implements InvocationHandler {
         @SuppressWarnings("unchecked")
         T t = (T) resource;
         return t;
-    }
-
-    /**
-     * Returns some of the more commonly used return types that are valid for a life cycle method.
-     *
-     * @param singularClassName        simple class name of the entity
-     * @param hasSingularEntityParam   if the life cycle method entity parameter is singular (not an Iterable or array)
-     * @param includeBooleanAndNumeric whether to include boolean and numeric types as valid.
-     * @return some of the more commonly used return types that are valid for a life cycle method.
-     */
-    private List<String> getValidReturnTypes(String singularClassName, boolean hasSingularEntityParam, boolean includeBooleanAndNumeric) {
-        List<String> validReturnTypes = new ArrayList<>();
-        if (includeBooleanAndNumeric) {
-            validReturnTypes.add("boolean");
-            validReturnTypes.add("int");
-            validReturnTypes.add("long");
-        }
-
-        validReturnTypes.add("void");
-
-        if (hasSingularEntityParam) {
-            validReturnTypes.add(singularClassName);
-        } else {
-            validReturnTypes.add(singularClassName + "[]");
-            validReturnTypes.add("List<" + singularClassName + ">");
-        }
-
-        return validReturnTypes;
-    }
-
-    /**
-     * Inserts entities (or records) into the database.
-     * An error is raised if any of the entities (or records) already exist in the database.
-     *
-     * @param arg       the entity or record, or array or Iterable or Stream of entity or record.
-     * @param queryInfo query information.
-     * @param em        the entity manager.
-     * @return the updated entities, using the return type that is required by Insert method signature.
-     * @throws Exception if an error occurs.
-     */
-    private Object insert(Object arg, QueryInfo queryInfo, EntityManager em) throws Exception {
-        boolean resultVoid = void.class.equals(queryInfo.singleType) ||
-                             Void.class.equals(queryInfo.singleType);
-        ArrayList<Object> results;
-
-        boolean hasSingularEntityParam = false;
-        if (queryInfo.entityParamType.isArray()) {
-            int length = Array.getLength(arg);
-            results = resultVoid ? null : new ArrayList<>(length);
-            for (int i = 0; i < length; i++) {
-                Object entity = queryInfo.toEntity(Array.get(arg, i));
-                em.persist(entity);
-                if (results != null)
-                    results.add(entity);
-            }
-            em.flush();
-        } else {
-            arg = arg instanceof Stream //
-                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
-                            : arg;
-
-            if (arg instanceof Iterable) {
-                results = resultVoid ? null : new ArrayList<>();
-                for (Object e : ((Iterable<?>) arg)) {
-                    Object entity = queryInfo.toEntity(e);
-                    em.persist(entity);
-                    if (results != null)
-                        results.add(entity);
-                }
-                em.flush();
-            } else {
-                hasSingularEntityParam = true;
-                results = resultVoid ? null : new ArrayList<>(1);
-                Object entity = queryInfo.toEntity(arg);
-                em.persist(entity);
-                em.flush();
-                if (results != null)
-                    results.add(entity);
-            }
-        }
-
-        Class<?> returnType = queryInfo.method.getReturnType();
-        Object returnValue;
-        if (resultVoid) {
-            returnValue = null;
-        } else {
-            if (queryInfo.entityInfo.recordClass != null)
-                for (int i = 0; i < results.size(); i++)
-                    results.set(i, queryInfo.entityInfo.toRecord(results.get(i)));
-
-            if (queryInfo.returnArrayType != null) {
-                Object[] newArray = (Object[]) Array.newInstance(queryInfo.returnArrayType, results.size());
-                returnValue = results.toArray(newArray);
-            } else {
-                Class<?> multiType = queryInfo.multiType;
-                if (multiType == null)
-                    if (results.size() == 1)
-                        returnValue = results.get(0);
-                    else if (results.isEmpty())
-                        returnValue = null;
-                    else
-                        throw excNonUniqueResult(queryInfo, results.size());
-                else if (multiType.isInstance(results))
-                    returnValue = results;
-                else if (Stream.class.equals(multiType))
-                    returnValue = results.stream();
-                else if (Iterable.class.isAssignableFrom(multiType))
-                    returnValue = queryInfo.convertToIterable(results, null, multiType);
-                else if (Iterator.class.equals(multiType))
-                    returnValue = results.iterator();
-                else
-                    throw exc(MappingException.class,
-                              "CWWKD1003.lifecycle.rtrn.err",
-                              queryInfo.method.getGenericReturnType().getTypeName(),
-                              queryInfo.method.getName(),
-                              repositoryInterface.getName(),
-                              "Insert",
-                              getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                                  hasSingularEntityParam,
-                                                  false));
-            }
-        }
-
-        if (CompletableFuture.class.equals(returnType) || CompletionStage.class.equals(returnType)) {
-            returnValue = CompletableFuture.completedFuture(returnValue); // useful for @Asynchronous
-        } else if (!resultVoid && !returnType.isInstance(returnValue)) {
-            throw exc(MappingException.class,
-                      "CWWKD1003.lifecycle.rtrn.err",
-                      queryInfo.method.getGenericReturnType().getTypeName(),
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      "Insert",
-                      getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                          hasSingularEntityParam,
-                                          false));
-        }
-
-        return returnValue;
     }
 
     @FFDCIgnore(Throwable.class)
@@ -993,12 +556,12 @@ public class RepositoryImpl<R> implements InvocationHandler {
                 switch (queryType) {
                     case SAVE: {
                         em = entityInfo.builder.createEntityManager();
-                        returnValue = save(args[0], queryInfo, em);
+                        returnValue = queryInfo.save(args[0], em);
                         break;
                     }
                     case INSERT: {
                         em = entityInfo.builder.createEntityManager();
-                        returnValue = insert(args[0], queryInfo, em);
+                        returnValue = queryInfo.insert(args[0], em);
                         break;
                     }
                     case FIND:
@@ -1263,8 +826,11 @@ public class RepositoryImpl<R> implements InvocationHandler {
                                         }
                                     if (firstNonNullResult == null
                                         || queryInfo.type == QueryInfo.Type.FIND_AND_DELETE
-                                        || queryInfo.returnArrayType != Object.class && queryInfo.returnArrayType.isInstance(firstNonNullResult)
-                                        || queryInfo.returnArrayType.isPrimitive() && isWrapperClassFor(queryInfo.returnArrayType, firstNonNullResult.getClass())) {
+                                        || queryInfo.returnArrayType != Object.class &&
+                                           queryInfo.returnArrayType.isInstance(firstNonNullResult)
+                                        || queryInfo.returnArrayType.isPrimitive() &&
+                                           Util.isWrapperClassFor(queryInfo.returnArrayType,
+                                                                   firstNonNullResult.getClass())) {
                                         returnValue = Array.newInstance(queryInfo.returnArrayType, size);
                                         int i = 0;
                                         for (Object result : results)
@@ -1314,7 +880,7 @@ public class RepositoryImpl<R> implements InvocationHandler {
                                         } else {
                                             // List<Object[]> with multiple Object[] elements
                                             // cannot convert to a one dimensional array
-                                            throw excNonUniqueResult(queryInfo, size);
+                                            throw queryInfo.excNonUniqueResult(size);
                                         }
                                     } else {
                                         throw exc(MappingException.class,
@@ -1378,15 +944,15 @@ public class RepositoryImpl<R> implements InvocationHandler {
                         if (arg instanceof Iterable) {
                             for (Object e : ((Iterable<?>) arg)) {
                                 numExpected++;
-                                updateCount += remove(e, queryInfo, em);
+                                updateCount += queryInfo.remove(e, em);
                             }
                         } else if (queryInfo.entityParamType.isArray()) {
                             numExpected = Array.getLength(arg);
                             for (int i = 0; i < numExpected; i++)
-                                updateCount += remove(Array.get(arg, i), queryInfo, em);
+                                updateCount += queryInfo.remove(Array.get(arg, i), em);
                         } else {
                             numExpected = 1;
-                            updateCount = remove(arg, queryInfo, em);
+                            updateCount = queryInfo.remove(arg, em);
                         }
 
                         if (updateCount < numExpected)
@@ -1422,15 +988,15 @@ public class RepositoryImpl<R> implements InvocationHandler {
                         if (arg instanceof Iterable) {
                             for (Object e : ((Iterable<?>) arg)) {
                                 numExpected++;
-                                updateCount += update(e, queryInfo, em);
+                                updateCount += queryInfo.update(e, em);
                             }
                         } else if (queryInfo.entityParamType.isArray()) {
                             numExpected = Array.getLength(arg);
                             for (int i = 0; i < numExpected; i++)
-                                updateCount += update(Array.get(arg, i), queryInfo, em);
+                                updateCount += queryInfo.update(Array.get(arg, i), em);
                         } else {
                             numExpected = 1;
-                            updateCount = update(arg, queryInfo, em);
+                            updateCount = queryInfo.update(arg, em);
                         }
 
                         if (updateCount < numExpected)
@@ -1456,12 +1022,12 @@ public class RepositoryImpl<R> implements InvocationHandler {
                     }
                     case UPDATE_WITH_ENTITY_PARAM_AND_RESULT: {
                         em = entityInfo.builder.createEntityManager();
-                        returnValue = findAndUpdate(args[0], queryInfo, em);
+                        returnValue = queryInfo.findAndUpdate(args[0], em);
                         break;
                     }
                     case COUNT: {
                         em = entityInfo.builder.createEntityManager();
-                        returnValue = count(em, queryInfo, args);
+                        returnValue = queryInfo.count(em, args);
                         break;
                     }
                     case EXISTS: {
@@ -1536,24 +1102,6 @@ public class RepositoryImpl<R> implements InvocationHandler {
     }
 
     /**
-     * Indicates if the specified class is a wrapper for the primitive class.
-     *
-     * @param primitive primitive class.
-     * @param cl        another class that might be a wrapper class for the primitive class.
-     * @return true if the class is the wrapper class for the primitive class, otherwise false.
-     */
-    private final boolean isWrapperClassFor(Class<?> primitive, Class<?> cl) {
-        return primitive == long.class && cl == Long.class ||
-               primitive == int.class && cl == Integer.class ||
-               primitive == float.class && cl == Float.class ||
-               primitive == double.class && cl == Double.class ||
-               primitive == char.class && cl == Character.class ||
-               primitive == byte.class && cl == Byte.class ||
-               primitive == boolean.class && cl == Boolean.class ||
-               primitive == short.class && cl == Short.class;
-    }
-
-    /**
      * Requires a single result.
      *
      * @param queryInfo information about the query.
@@ -1570,107 +1118,7 @@ public class RepositoryImpl<R> implements InvocationHandler {
         else if (size == 0)
             throw excEmptyResult(queryInfo.method);
         else
-            throw excNonUniqueResult(queryInfo, results.size());
-    }
-
-    /**
-     * Removes the entity (or record) from the database if its attributes match the database.
-     *
-     * @param e         the entity or record.
-     * @param queryInfo query information that is prepopulated for deleteById.
-     * @param em        the entity manager.
-     * @return the number of entities deleted (1 or 0).
-     * @throws Exception if an error occurs or if the repository method return type is void and
-     *                       the entity (or correct version of the entity) was not found.
-     */
-    private int remove(Object e, QueryInfo queryInfo, EntityManager em) throws Exception {
-        Class<?> entityClass = queryInfo.entityInfo.getType();
-
-        if (e == null)
-            throw exc(NullPointerException.class,
-                      "CWWKD1015.null.entity.param",
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName());
-
-        if (!entityClass.isInstance(e))
-            throw exc(IllegalArgumentException.class,
-                      "CWWKD1016.incompat.entity.param",
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      entityClass.getName(),
-                      e.getClass().getName());
-
-        EntityInfo entityInfo = queryInfo.entityInfo;
-        String jpql = queryInfo.jpql;
-
-        int versionParamIndex = (entityInfo.idClassAttributeAccessors == null ? 1 : entityInfo.idClassAttributeAccessors.size()) + 1;
-        Object version = null;
-        if (entityInfo.versionAttributeName != null) {
-            version = queryInfo.getAttribute(e, entityInfo.versionAttributeName);
-            if (version == null)
-                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
-        }
-
-        Object id = null;
-        String idAttributeName = null;
-        if (entityInfo.idClassAttributeAccessors == null) {
-            idAttributeName = queryInfo.getAttributeName(ID, true);
-            id = queryInfo.getAttribute(e, idAttributeName);
-            if (id == null) {
-                jpql = jpql.replace("=?" + (versionParamIndex - 1), " IS NULL");
-                if (version != null)
-                    jpql = jpql.replace("=?" + versionParamIndex, "=?" + (versionParamIndex - 1));
-            }
-        }
-
-        if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled() && jpql != queryInfo.jpql)
-            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
-
-        TypedQuery<?> delete = em.createQuery(jpql, entityInfo.entityClass);
-
-        if (entityInfo.idClassAttributeAccessors == null) {
-            int p = 1;
-            if (id != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(id));
-                delete.setParameter(p++, id);
-            }
-            if (version != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(version));
-                delete.setParameter(p, version);
-            }
-        } else {
-            queryInfo.setParametersFromIdClassAndVersion(1, delete, e, version);
-        }
-
-        int numDeleted = delete.executeUpdate();
-
-        if (numDeleted == 0) {
-            Class<?> returnType = queryInfo.method.getReturnType();
-            if (void.class.equals(returnType) || Void.class.equals(returnType)) {
-                if (idAttributeName == null)
-                    idAttributeName = ID;
-                List<String> entityProps = new ArrayList<>(2);
-                if (id != null)
-                    entityProps.add(queryInfo.loggableAppend(idAttributeName,
-                                                             "=", id));
-                if (entityInfo.versionAttributeName != null && version != null)
-                    entityProps.add(queryInfo.loggableAppend(entityInfo.versionAttributeName,
-                                                             "=", version));
-                throw exc(OptimisticLockingFailureException.class,
-                          "CWWKD1050.opt.lock.exc",
-                          queryInfo.method.getName(),
-                          repositoryInterface.getName(),
-                          e.getClass().getName(),
-                          entityProps,
-                          LIFE_CYCLE_METHODS_THAT_RETURN_ENTITIES);
-            }
-        } else if (numDeleted > 1) {
-            throw new DataException("Found " + numDeleted + " matching entities."); // ought to be unreachable
-        }
-
-        return numDeleted;
+            throw queryInfo.excNonUniqueResult(results.size());
     }
 
     /**
@@ -1701,193 +1149,5 @@ public class RepositoryImpl<R> implements InvocationHandler {
                       repositoryInterface.getName(),
                       queryInfo.type == Type.DELETE ? "Delete" : "Update");
         return result;
-    }
-
-    /**
-     * Saves entities (or records) to the database, which can involve an update or an insert,
-     * depending on whether the entity already exists.
-     *
-     * @param arg       the entity or record, or array or Iterable or Stream of entity or record.
-     * @param queryInfo query information.
-     * @param em        the entity manager.
-     * @return the updated entities, using the return type that is required by the repository Save method signature.
-     * @throws Exception if an error occurs.
-     */
-    private Object save(Object arg, QueryInfo queryInfo, EntityManager em) throws Exception {
-        boolean resultVoid = void.class.equals(queryInfo.singleType) ||
-                             Void.class.equals(queryInfo.singleType);
-        List<Object> results;
-
-        boolean hasSingularEntityParam = false;
-        if (queryInfo.entityParamType.isArray()) {
-            results = new ArrayList<>();
-            int length = Array.getLength(arg);
-            for (int i = 0; i < length; i++)
-                results.add(em.merge(queryInfo.toEntity(Array.get(arg, i))));
-            em.flush();
-        } else {
-            arg = arg instanceof Stream //
-                            ? ((Stream<?>) arg).sequential().collect(Collectors.toList()) //
-                            : arg;
-
-            if (Iterable.class.isAssignableFrom(queryInfo.entityParamType)) {
-                results = new ArrayList<>();
-                for (Object e : ((Iterable<?>) arg))
-                    results.add(em.merge(queryInfo.toEntity(e)));
-                em.flush();
-            } else {
-                hasSingularEntityParam = true;
-                results = resultVoid ? null : new ArrayList<>(1);
-                Object entity = em.merge(queryInfo.toEntity(arg));
-                if (results != null)
-                    results.add(entity);
-                em.flush();
-            }
-        }
-
-        Class<?> returnType = queryInfo.method.getReturnType();
-        Object returnValue;
-        if (resultVoid) {
-            returnValue = null;
-        } else {
-            if (queryInfo.entityInfo.recordClass != null)
-                for (int i = 0; i < results.size(); i++)
-                    results.set(i, queryInfo.entityInfo.toRecord(results.get(i)));
-
-            if (queryInfo.returnArrayType != null) {
-                Object[] newArray = (Object[]) Array.newInstance(queryInfo.returnArrayType, results.size());
-                returnValue = results.toArray(newArray);
-            } else {
-                Class<?> multiType = queryInfo.multiType;
-                if (multiType == null)
-                    if (results.size() == 1)
-                        returnValue = results.get(0);
-                    else if (results.isEmpty())
-                        returnValue = null;
-                    else
-                        throw excNonUniqueResult(queryInfo, results.size());
-                else if (multiType.isInstance(results))
-                    returnValue = results;
-                else if (Stream.class.equals(multiType))
-                    returnValue = results.stream();
-                else if (Iterable.class.isAssignableFrom(multiType))
-                    returnValue = queryInfo.convertToIterable(results, null, multiType);
-                else if (Iterator.class.equals(multiType))
-                    returnValue = results.iterator();
-                else
-                    throw exc(MappingException.class,
-                              "CWWKD1003.lifecycle.rtrn.err",
-                              queryInfo.method.getGenericReturnType().getTypeName(),
-                              queryInfo.method.getName(),
-                              repositoryInterface.getName(),
-                              "Save",
-                              getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                                  hasSingularEntityParam,
-                                                  false));
-            }
-        }
-
-        if (CompletableFuture.class.equals(returnType) || CompletionStage.class.equals(returnType)) {
-            returnValue = CompletableFuture.completedFuture(returnValue); // useful for @Asynchronous
-        } else if (!resultVoid && !returnType.isInstance(returnValue)) {
-            throw exc(MappingException.class,
-                      "CWWKD1003.lifecycle.rtrn.err",
-                      queryInfo.method.getGenericReturnType().getTypeName(),
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      "Save",
-                      getValidReturnTypes(results.get(0).getClass().getSimpleName(),
-                                          hasSingularEntityParam,
-                                          false));
-        }
-
-        return returnValue;
-    }
-
-    /**
-     * Updates the entity (or record) from the database if its attributes match the database.
-     *
-     * @param e         the entity or record.
-     * @param queryInfo query information that is prepopulated for update by id (and possibly version).
-     * @param em        the entity manager.
-     * @return the number of entities updated (1 or 0).
-     * @throws Exception if an error occurs.
-     */
-    private int update(Object e, QueryInfo queryInfo, EntityManager em) throws Exception {
-        Class<?> entityClass = queryInfo.entityInfo.getType();
-
-        if (e == null)
-            throw exc(NullPointerException.class,
-                      "CWWKD1015.null.entity.param",
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName());
-
-        if (!entityClass.isInstance(e))
-            throw exc(IllegalArgumentException.class,
-                      "CWWKD1016.incompat.entity.param",
-                      queryInfo.method.getName(),
-                      repositoryInterface.getName(),
-                      entityClass.getName(),
-                      e.getClass().getName());
-
-        String jpql = queryInfo.jpql;
-        EntityInfo entityInfo = queryInfo.entityInfo;
-        Set<String> attrsToUpdate = entityInfo.attributeNamesForEntityUpdate;
-
-        int versionParamIndex = entityInfo.idClassAttributeAccessors == null //
-                        ? (attrsToUpdate.size() + 2) //
-                        : (attrsToUpdate.size() +
-                           entityInfo.idClassAttributeAccessors.size() + 1);
-        Object version = null;
-        if (entityInfo.versionAttributeName != null) {
-            version = queryInfo.getAttribute(e, entityInfo.versionAttributeName);
-            if (version == null)
-                jpql = jpql.replace("=?" + versionParamIndex, " IS NULL");
-        }
-
-        Object id = null;
-        String idAttributeName = null;
-        if (entityInfo.idClassAttributeAccessors == null) {
-            idAttributeName = entityInfo.attributeNames.get(ID);
-            id = queryInfo.getAttribute(e, idAttributeName);
-            if (id == null) {
-                jpql = jpql.replace("=?" + (versionParamIndex - 1), " IS NULL");
-                if (version != null)
-                    jpql = jpql.replace("=?" + versionParamIndex, "=?" + (versionParamIndex - 1));
-            }
-        }
-
-        if (TraceComponent.isAnyTracingEnabled() && jpql != queryInfo.jpql)
-            Tr.debug(this, tc, "JPQL adjusted for NULL id or version", jpql);
-
-        TypedQuery<?> update = em.createQuery(jpql, entityInfo.entityClass);
-
-        // parameters for entity attributes to update:
-        int p = 1;
-        for (String attrName : attrsToUpdate)
-            queryInfo.setParameter(p++, update, e, attrName);
-
-        // id and version parameters
-        if (entityInfo.idClassAttributeAccessors == null) {
-            if (id != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(id));
-                update.setParameter(p++, id);
-            }
-
-            if (entityInfo.versionAttributeName != null && version != null) {
-                if (TraceComponent.isAnyTracingEnabled() && tc.isDebugEnabled())
-                    Tr.debug(tc, "set ?" + p + ' ' + queryInfo.loggable(version));
-                update.setParameter(p++, version);
-            }
-        } else { // has IdClass
-            queryInfo.setParametersFromIdClassAndVersion(p, update, e, version);
-        }
-
-        int numUpdated = update.executeUpdate();
-
-        if (numUpdated > 1)
-            throw new DataException("Found " + numUpdated + " matching entities."); // ought to be unreachable
-        return numUpdated;
     }
 }
